move NestedRoute to components folder root
