+++ index.js
@@ -12,4 +12,4 @@ export { default as HandlerManager } from './src/components/HandlerManager';
 export { default as coreEvents } from './src/events';
 export { default as IntlConsumer } from './src/components/IntlConsumer';
 export { default as AppIcon } from './src/components/AppIcon';
-export { Route, Switch, Redirect } from './src/components/Router/Route';
+export { Route, Switch, Redirect } from './src/components/NestedRouter';
